+++ examples/chat-vue/CHANGELOG.md
@@ -1,5 +1,23 @@
 # chat-vue
 
+## 0.0.91
+
+### Patch Changes
+
+- Updated dependencies [439f0fe]
+  - jazz-tools@0.13.20
+  - jazz-browser@0.13.20
+  - jazz-vue@0.13.20
+
+## 0.0.90
+
+### Patch Changes
+
+- Updated dependencies [80530a4]
+  - jazz-tools@0.13.19
+  - jazz-browser@0.13.19
+  - jazz-vue@0.13.19
+
 ## 0.0.89
 
 ### Patch Changes