+++ examples/chat/CHANGELOG.md
@@ -1,5 +1,23 @@
 # jazz-example-chat
 
+## 0.0.189
+
+### Patch Changes
+
+- Updated dependencies [439f0fe]
+  - jazz-tools@0.13.20
+  - jazz-inspector@0.13.20
+  - jazz-react@0.13.20
+
+## 0.0.188
+
+### Patch Changes
+
+- Updated dependencies [80530a4]
+  - jazz-tools@0.13.19
+  - jazz-inspector@0.13.19
+  - jazz-react@0.13.19
+
 ## 0.0.187
 
 ### Patch Changes