+++ examples/form/CHANGELOG.md
@@ -1,5 +1,21 @@
 # form
 
+## 0.1.29
+
+### Patch Changes
+
+- Updated dependencies [439f0fe]
+  - jazz-tools@0.13.20
+  - jazz-react@0.13.20
+
+## 0.1.28
+
+### Patch Changes
+
+- Updated dependencies [80530a4]
+  - jazz-tools@0.13.19
+  - jazz-react@0.13.19
+
 ## 0.1.27
 
 ### Patch Changes