+++ examples/image-upload/CHANGELOG.md
@@ -1,5 +1,21 @@
 # image-upload
 
+## 0.0.85
+
+### Patch Changes
+
+- Updated dependencies [439f0fe]
+  - jazz-tools@0.13.20
+  - jazz-react@0.13.20
+
+## 0.0.84
+
+### Patch Changes
+
+- Updated dependencies [80530a4]
+  - jazz-tools@0.13.19
+  - jazz-react@0.13.19
+
 ## 0.0.83
 
 ### Patch Changes