+++ examples/pets/CHANGELOG.md
@@ -1,5 +1,21 @@
 # jazz-example-pets
 
+## 0.0.205
+
+### Patch Changes
+
+- Updated dependencies [439f0fe]
+  - jazz-tools@0.13.20
+  - jazz-react@0.13.20
+
+## 0.0.204
+
+### Patch Changes
+
+- Updated dependencies [80530a4]
+  - jazz-tools@0.13.19
+  - jazz-react@0.13.19
+
 ## 0.0.203
 
 ### Patch Changes