+++ packages/jazz-browser-media-images/CHANGELOG.md
@@ -1,5 +1,21 @@
 # jazz-browser-media-images
 
+## 0.13.20
+
+### Patch Changes
+
+- Updated dependencies [439f0fe]
+  - jazz-tools@0.13.20
+  - jazz-browser@0.13.20
+
+## 0.13.19
+
+### Patch Changes
+
+- Updated dependencies [80530a4]
+  - jazz-tools@0.13.19
+  - jazz-browser@0.13.19
+
 ## 0.13.18
 
 ### Patch Changes