+++ packages/jazz-richtext-prosemirror/CHANGELOG.md
@@ -1,5 +1,21 @@
 # jazz-richtext-prosemirror
 
+## 0.1.9
+
+### Patch Changes
+
+- Updated dependencies [439f0fe]
+  - jazz-tools@0.13.20
+  - jazz-browser@0.13.20
+
+## 0.1.8
+
+### Patch Changes
+
+- Updated dependencies [80530a4]
+  - jazz-tools@0.13.19
+  - jazz-browser@0.13.19
+
 ## 0.1.7
 
 ### Patch Changes